Add first batch of cases.
This is the first 150 (out of 1,052 at this time). This is manually created metadata that given a NYPD discipline document URL, contains the tax IDs and disciplinary case numbers associated with it. Later commits will hace data for the remaining cases.

I still need to do some verification (checking to make sure tax IDs and case numbers match a regular expression, and verifying that the info the metadata shows up in the DocumentCloud OCR version of the document) before using this, because typos are likely.