Follow up to previous commit: add a list of bad documents to skip.
The previous APU report accidentally linked to the wrong docs, but docs that were still in the report so still got uploaded (and thus wouldn't try to get uploaded again). These doc links are just wrong and don't exist, so this needs to skip them so they don't get uploaded again every time.